WIP : Fix event listening
