--- .snyk
@@ -1,8 +1,0 @@
-# Snyk (https://snyk.io) policy file, patches or ignores known vulnerabilities.
-version: v1.13.3
-ignore: {}
-# patches apply the minimum changes required to fix a vulnerability
-patch:
-  SNYK-JS-AXIOS-174505:
-    - '@slack/client > axios':
-        patched: '2019-05-06T02:02:58.021Z'
