--- .travis.yml
@@ -1,12 +1,0 @@
-language: node_js
-node_js:
-- node
-script:
-- npm install
-before_install:
-- openssl aes-256-cbc -K $encrypted_a12d32dc32b7_key -iv $encrypted_a12d32dc32b7_iv -in ./.travis/id_rsa.enc -out ./.travis/id_rsa -d
-after_success:
-- ".travis/deploy.sh"
-addons:
-  ssh_known_hosts:
-  - 52.47.181.19
