+++ src/slack.js
@@ -1,90 +1,16 @@
-const { App, LogLevel } = require('@slack/bolt');
 const schedule = require("node-schedule");
-const db = require("./db.js");
-const dialogs = require("./dialogs.js");
 const logger = require("./logger.js");
-const workspaces = require("./workspaces.js");
 
-const HttpsProxyAgent = require('https-proxy-agent');
-const proxy = process.env.HTTP_PROXY ? new HttpsProxyAgent(process.env.HTTP_PROXY) : null;
-var app;
+let app
 
 var initJobs = function () {
     schedule.scheduleJob("* * * * * *", postShift);
     schedule.scheduleJob("*/2 * * * * *", updateShift);
 };
 
-const authorizeFn = async ({ teamId }) => {
-    var marchWorkspaceId = { team_id: teamId };
-    db.read("workspaces", marchWorkspaceId, function(workspace) {
-        if(workspace != null) {
-            return {
-                botToken: workspace.access_token,
-                botId: workspace.bot_id,
-                botUserId: workspace.bot_user_id
-            };
-        } else {
-            logger.log('No matching authorizations');
-        }
-    });
+var setApp = function(newApp) {
+    app = newApp;
 }
-
-var initApp = function (receiver) {
-    app = new App({
-        signingSecret: process.env.SLACK_SIGNING_SECRET,
-        agent: proxy,
-        authorize: authorizeFn,
-        logLevel: LogLevel.INFO,
-        receiver: receiver
-    });
-    
-    app.message(async ({ message }) => {
-        if (message.text !== undefined) {
-            db.insert("messages", message);
-        }
-        if (message.text === ":house:") {
-            db.read("workspaces", { team_id: message.team }, function (workspacesInDb) {
-                var workspace = workspacesInDb;
-                if (workspacesInDb.access_token === undefined) {
-                    workspace = workspacesInDb[0];
-                }
-                var user = workspaces.getUsersByChannelId(workspace, message.channel);
-                if (user !== null) {
-                    db.read("dialogs", { name: "Consent PM" }, function (dialog) {
-                        dialog.channelId = message.channel;
-                        dialogs.speakRecurse(workspace, dialog, "0", () => { });
-                    })
-                }
-            });
-        }
-    });
-
-    app.event('team_join', async ({ event }) => {
-        db.read("workspaces", { team_id: event.user.team_id }, function (workspacesOfNewUser) {
-            if (!Array.isArray(workspacesOfNewUser)) {
-                workspacesOfNewUser = [workspacesOfNewUser];
-            }
-            var previous_bot_access_token = [];
-            for (var workspaceNum in workspacesOfNewUser) {
-                var workspaceOfNewUser = workspacesOfNewUser[workspaceNum];
-                if (previous_bot_access_token.indexOf(workspaceOfNewUser.access_token) < 0) {
-                    workspaces.openIM(workspaceOfNewUser, [event.user], 0, function () {
-                        var workspaceId = workspaceOfNewUser._id;
-                        db.update("workspaces", { _id: new db.mongodb().ObjectId(workspaceId) }, workspaceOfNewUser, function () {
-                            db.read("dialogs", { name: "Consent PM" }, function (dialog) {
-                                dialog.channelId = workspaces.getUsersById(workspaceOfNewUser, event.user.id).im_id;
-                                workspaceOfNewUser._id = workspaceId;
-                                dialogs.speakRecurse(workspaceOfNewUser, dialog, 0);
-                            });
-                        });
-                    });
-                    previous_bot_access_token.push(workspaceOfNewUser.access_token);
-                }
-            }
-        });
-    });
-    return app;
-};
 
 var authTest = function (workspace) {
     return app.client.auth.test({ token: workspace.access_token });
@@ -95,7 +21,7 @@ var join = function (workspace, channelId) {
 };
 
 var listUsers = function (workspace) {
-    return app.client.users.list({ token: workspace.access_token});
+    return app.client.users.list({ token: workspace.access_token });
 };
 
 var openIM = function (workspace, params) {
@@ -177,7 +103,7 @@ var uploadFiles = function (workspace, files) {
     return app.client.files.upload(files);
 };
 
-exports.initApp = initApp;
+exports.setApp = setApp;
 exports.initJobs = initJobs;
 exports.authTest = authTest;
 exports.join = join;
